support update policy

+++ pkg/manager/member/pd_upgrader.go
@@ -17,6 +17,8 @@ import (
 	"fmt"
 	"sort"
 
+	mngerutils "github.com/pingcap/tidb-operator/pkg/manager/utils"
+
 	"github.com/pingcap/advanced-statefulset/client/apis/apps/v1/helper"
 	"github.com/pingcap/errors"
 	"github.com/pingcap/tidb-operator/pkg/apis/pingcap/v1alpha1"
@@ -43,7 +45,6 @@ func (u *pdUpgrader) Upgrade(tc *v1alpha1.TidbCluster, oldSet *apps.StatefulSet,
 }
 
 func (u *pdUpgrader) gracefulUpgrade(tc *v1alpha1.TidbCluster, oldSet *apps.StatefulSet, newSet *apps.StatefulSet) error {
-
 	ns := tc.GetNamespace()
 	tcName := tc.GetName()
 	if !tc.Status.PD.Synced {
@@ -68,37 +69,75 @@ func (u *pdUpgrader) gracefulUpgrade(tc *v1alpha1.TidbCluster, oldSet *apps.Stat
 	if tc.Status.PD.StatefulSet.UpdateRevision == tc.Status.PD.StatefulSet.CurrentRevision {
 		return nil
 	}
-	pods, err := u.podsToUpgrade(tc, oldSet)
-	if err != nil {
-		return err
-	}
-	for _, pod := range pods {
-		podName := pod.Name
-		pod1Ordinal, err := util.GetOrdinalFromPodName(podName)
+	if tc.Spec.PD.IsEnableIntelligentOperation != nil && *tc.Spec.PD.IsEnableIntelligentOperation {
+		// get need to upgrade pod list
+		pods, err := u.podsToUpgrade(tc, oldSet)
 		if err != nil {
-			continue
+			return err
 		}
-		pod, err := u.deps.PodLister.Pods(ns).Get(podName)
-		if err != nil {
-			return fmt.Errorf("gracefulUpgrade: failed to get pods %s for cluster %s/%s, error: %s", podName, ns, tcName, err)
-		}
-
-		revision, exist := pod.Labels[apps.ControllerRevisionHashLabelKey]
-		if !exist {
-			return controller.RequeueErrorf("tidbcluster: [%s/%s]'s pd pod: [%s] has no label: %s", ns, tcName, podName, apps.ControllerRevisionHashLabelKey)
-		}
-
-		if revision == tc.Status.PD.StatefulSet.UpdateRevision {
-			if member, exist := tc.Status.PD.Members[PdName(tc.Name, pod1Ordinal, tc.Namespace, tc.Spec.ClusterDomain)]; !exist || !member.Health {
-				return controller.RequeueErrorf("tidbcluster: [%s/%s]'s pd upgraded pod: [%s] is not ready", ns, tcName, podName)
+		// sort candidates by pd leader or no-leader
+		sortedCandidates := u.sortCandidates(tc, pods)
+		for _, candidate := range sortedCandidates {
+			podName := candidate.Name
+			pod1Ordinal, err := util.GetOrdinalFromPodName(podName)
+			if err != nil {
+				continue
 			}
-			continue
-		}
-		klog.Infof("cwtest upgradePod:%s", pod.Name)
-		u.deps.Recorder.Event(tc, corev1.EventTypeNormal, fmt.Sprintf("PDUpgrade"), pod.Name)
-		return u.upgradePDPod(tc, pod1Ordinal, newSet)
-	}
+			pod, err := u.deps.PodLister.Pods(ns).Get(podName)
+			if err != nil {
+				return fmt.Errorf("gracefulUpgrade: failed to get pods %s for cluster %s/%s, error: %s", podName, ns, tcName, err)
+			}
 
+			revision, exist := pod.Labels[apps.ControllerRevisionHashLabelKey]
+			if !exist {
+				return controller.RequeueErrorf("tidbcluster: [%s/%s]'s pd pod: [%s] has no label: %s", ns, tcName, podName, apps.ControllerRevisionHashLabelKey)
+			}
+
+			if revision == tc.Status.PD.StatefulSet.UpdateRevision {
+				if member, exist := tc.Status.PD.Members[PdName(tc.Name, pod1Ordinal, tc.Namespace, tc.Spec.ClusterDomain)]; !exist || !member.Health {
+					return controller.RequeueErrorf("tidbcluster: [%s/%s]'s pd upgraded pod: [%s] is not ready", ns, tcName, podName)
+				}
+				continue
+			}
+			u.deps.Recorder.Event(tc, corev1.EventTypeNormal, fmt.Sprintf("PDUpgrade"), fmt.Sprintf("%s:upgrade pod %s", SplitRevision(tc.Status.PD.StatefulSet.UpdateRevision), pod.Name))
+			return u.upgradePDPod(tc, pod1Ordinal, newSet)
+		}
+	} else {
+		if oldSet.Spec.UpdateStrategy.Type == apps.OnDeleteStatefulSetStrategyType || oldSet.Spec.UpdateStrategy.RollingUpdate == nil {
+			// Manually bypass tidb-operator to modify statefulset directly, such as modify pd statefulset's RollingUpdate straregy to OnDelete strategy,
+			// or set RollingUpdate to nil, skip tidb-operator's rolling update logic in order to speed up the upgrade in the test environment occasionally.
+			// If we encounter this situation, we will let the native statefulset controller do the upgrade completely, which may be unsafe for upgrading pd.
+			// Therefore, in the production environment, we should try to avoid modifying the pd statefulset update strategy directly.
+			newSet.Spec.UpdateStrategy = oldSet.Spec.UpdateStrategy
+			klog.Warningf("tidbcluster: [%s/%s] pd statefulset %s UpdateStrategy has been modified manually", ns, tcName, oldSet.GetName())
+			return nil
+		}
+
+		mngerutils.SetUpgradePartition(newSet, *oldSet.Spec.UpdateStrategy.RollingUpdate.Partition)
+		podOrdinals := helper.GetPodOrdinals(*oldSet.Spec.Replicas, oldSet).List()
+		for _i := len(podOrdinals) - 1; _i >= 0; _i-- {
+			i := podOrdinals[_i]
+			podName := PdPodName(tcName, i)
+			pod, err := u.deps.PodLister.Pods(ns).Get(podName)
+			if err != nil {
+				return fmt.Errorf("gracefulUpgrade: failed to get pods %s for cluster %s/%s, error: %s", podName, ns, tcName, err)
+			}
+
+			revision, exist := pod.Labels[apps.ControllerRevisionHashLabelKey]
+			if !exist {
+				return controller.RequeueErrorf("tidbcluster: [%s/%s]'s pd pod: [%s] has no label: %s", ns, tcName, podName, apps.ControllerRevisionHashLabelKey)
+			}
+
+			if revision == tc.Status.PD.StatefulSet.UpdateRevision {
+				if member, exist := tc.Status.PD.Members[PdName(tc.Name, i, tc.Namespace, tc.Spec.ClusterDomain)]; !exist || !member.Health {
+					return controller.RequeueErrorf("tidbcluster: [%s/%s]'s pd upgraded pod: [%s] is not ready", ns, tcName, podName)
+				}
+				continue
+			}
+
+			return u.upgradePDPod(tc, i, newSet)
+		}
+	}
 	return nil
 }
 
@@ -132,18 +171,14 @@ func (u *pdUpgrader) podsToUpgrade(tc *v1alpha1.TidbCluster, set *apps.StatefulS
 		}
 	}
 
-	// Step 1. Sort the Pods to get the ones with the higher priority
-	candidates := make([]*corev1.Pod, len(toUpgrade))
-	copy(candidates, toUpgrade)
-	sortCandidates(tc, candidates)
-	for index, can := range candidates {
-		klog.Infof("cwtest candidates:%d,%s", index, can.Name)
-	}
-	return candidates, nil
+	return toUpgrade, nil
 }
 
-func sortCandidates(tc *v1alpha1.TidbCluster, allPods []*corev1.Pod) {
-	sort.Slice(allPods, func(i, j int) bool {
+func (u *pdUpgrader) sortCandidates(tc *v1alpha1.TidbCluster, allPods []*corev1.Pod) []*corev1.Pod {
+	// Step 1. Sort the Pods to get the ones with the higher priority
+	candidates := make([]*corev1.Pod, len(allPods))
+	copy(candidates, allPods)
+	sort.Slice(candidates, func(i, j int) bool {
 		pod1 := allPods[i]
 		pod2 := allPods[j]
 		// check if either is a master node. masters come after all other roles
@@ -172,6 +207,7 @@ func sortCandidates(tc *v1alpha1.TidbCluster, allPods []*corev1.Pod) {
 		}
 		return pod1PdName > pod2PdName
 	})
+	return candidates
 }
 
 func (u *pdUpgrader) upgradePDPod(tc *v1alpha1.TidbCluster, ordinal int32, newSet *apps.StatefulSet) error {
@@ -202,6 +238,7 @@ func (u *pdUpgrader) upgradePDPod(tc *v1alpha1.TidbCluster, ordinal int32, newSe
 			}
 		}
 		if len(targetName) > 0 {
+			u.deps.Recorder.Event(tc, corev1.EventTypeNormal, fmt.Sprintf("PDUpgrade"), fmt.Sprintf("%s:leader transfer to %s", SplitRevision(tc.Status.PD.StatefulSet.UpdateRevision), targetName))
 			err := u.transferPDLeaderTo(tc, targetName)
 			if err != nil {
 				klog.Errorf("pd upgrader: failed to transfer pd leader to: %s, %v", targetName, err)
@@ -212,19 +249,22 @@ func (u *pdUpgrader) upgradePDPod(tc *v1alpha1.TidbCluster, ordinal int32, newSe
 			return controller.RequeueErrorf("tidbcluster: [%s/%s]'s pd member: [%s] is transferring leader to pd member: [%s]", ns, tcName, upgradePdName, targetName)
 		}
 	}
-
-	pod, err := u.deps.PodLister.Pods(ns).Get(upgradePodName)
-	if err != nil && !errors.IsNotFound(err) {
-		return fmt.Errorf("pd upgrade[get pod]: failed to upgrade pod %s/%s for tc %s/%s, error: %s", ns, upgradePodName, ns, tcName, err)
-	}
-	if pod != nil {
-		if pod.DeletionTimestamp == nil {
-			if err := u.deps.PodControl.DeletePod(tc, pod); err != nil {
-				return err
+	if tc.Spec.PD.IsEnableIntelligentOperation != nil && *tc.Spec.PD.IsEnableIntelligentOperation {
+		pod, err := u.deps.PodLister.Pods(ns).Get(upgradePodName)
+		if err != nil && !errors.IsNotFound(err) {
+			return fmt.Errorf("pd upgrade[get pod]: failed to upgrade pod %s/%s for tc %s/%s, error: %s", ns, upgradePodName, ns, tcName, err)
+		}
+		if pod != nil {
+			if pod.DeletionTimestamp == nil {
+				if err := u.deps.PodControl.DeletePod(tc, pod); err != nil {
+					return err
+				}
 			}
+		} else {
+			klog.Infof("pd upgrade: get pod %s/%s not found, skip", ns, upgradePodName)
 		}
 	} else {
-		klog.Infof("pd upgrade: get pod %s/%s not found, skip", ns, upgradePodName)
+		mngerutils.SetUpgradePartition(newSet, ordinal)
 	}
 	return nil
 }
